+++ src/app/(landing)/transportation_carbon_footprint_calculator/page.tsx
@@ -4,7 +4,8 @@ import { useState, useRef } from 'react';
 import Head from 'next/head';
 import { Truck, Ship, Plane, MapPin, Leaf, ArrowRight, Loader2, Weight, Activity, Settings2, ChevronDown, ChevronUp, Sparkles } from 'lucide-react';
 import MapViewer from '@/components/map_viewer';
-import { calculateLogisticsPlan, ILogisticsPlan } from '@/lib/actions/route';
+import { calculateLogisticsPlan } from '@/lib/actions/route';
+import { ILogisticsPlan } from '@/interfaces/logistics';
 import { parseSmartInput } from '@/lib/actions/smart';
 // import * as htmlToImage from 'html-to-image';
 // import { jsPDF } from 'jspdf';
