+++ src/lib/actions/route.ts
@@ -1,47 +1,10 @@
 'use server';
 
-import { getNearestPort, getNearestAirport, INearestPortResult } from '@/lib/actions/logistics';
-import type { Geometry } from 'geojson';
+import { getNearestPort, getNearestAirport } from '@/lib/actions/logistics';
 import searoute from 'searoute-js';
 import { logisticsRepo } from "@/repositories/logistics.repo";
 import { esgRepo } from "@/repositories/esg.repo";
-// import { Prisma } from "@/generated/client";
-
-export interface ITransportSegment {
-    success: boolean;
-    distanceKm?: number;
-    distanceNm?: number;
-    co2eKg?: number;
-    geometry: Geometry | null;
-    isFallback?: boolean;
-}
-
-export interface IComparePlansResponse {
-    success: boolean;
-    plans: {
-        landOnly: ITransportSegment & { co2eKg?: number };
-        sea_multimodal: {
-            land_origin_to_port: ITransportSegment;
-            sea_port_to_port: ITransportSegment;
-            land_port_to_dest: ITransportSegment;
-            total_co2eKg?: number;
-        };
-        air_multimodal: {
-            land_origin_to_airport: ITransportSegment;
-            air_airport_to_airport: ITransportSegment;
-            land_airport_to_dest: ITransportSegment;
-            total_co2eKg?: number;
-        };
-    };
-}
-
-export interface ILogisticsPlan {
-    exportPort: INearestPortResult;
-    importPort: INearestPortResult;
-    exportAirport: INearestPortResult;
-    importAirport: INearestPortResult;
-    comparisonData: IComparePlansResponse;
-}
+import { ILogisticsPlan, ITransportSegment } from '@/interfaces/logistics';
 
 function calculateDistanceKm(lat1: number, lon1: number, lat2: number, lon2: number) {
     const R = 6371;
